refactor: extract service
