+++ src/cli/cmd-login.ts
@@ -1,11 +1,5 @@
 import { AuthenticationError, NpmLoginService } from "../services/npm-login";
-import {
-  GetUpmConfigDir,
-  GetUpmConfigDirError,
-  LoadUpmConfig,
-  tryStoreUpmAuth,
-  UpmAuthStoreError,
-} from "../io/upm-config-io";
+import { GetUpmConfigDir, GetUpmConfigDirError } from "../io/upm-config-io";
 import { EnvParseError, ParseEnvService } from "../services/parse-env";
 import { BasicAuth, encodeBasicAuth, TokenAuth } from "../domain/upm-config";
 import { coerceRegistryUrl } from "../domain/registry-url";
@@ -20,7 +14,7 @@ import { Ok, Result } from "ts-results-es";
 import { NpmrcLoadError, NpmrcSaveError } from "../io/npmrc-io";
 import { AuthNpmrcService } from "../services/npmrc-auth";
 import { Logger } from "npmlog";
-import { WriteTextFile } from "../io/file-io";
+import { SaveAuthToUpmConfig, UpmAuthStoreError } from "../services/upm-auth";
 
 /**
  * Errors which may occur when logging in.
@@ -61,9 +55,8 @@ export function makeLoginCmd(
   parseEnv: ParseEnvService,
   authNpmrc: AuthNpmrcService,
   npmLogin: NpmLoginService,
-  writeFile: WriteTextFile,
   getUpmConfigDir: GetUpmConfigDir,
-  loadUpmConfig: LoadUpmConfig,
+  saveAuthToUpmConfig: SaveAuthToUpmConfig,
   log: Logger
 ): LoginCmd {
   return async (options) => {
@@ -92,17 +85,11 @@ export function makeLoginCmd(
     if (options.basicAuth) {
       // basic auth
       const _auth = encodeBasicAuth(username, password);
-      const result = await tryStoreUpmAuth(
-        loadUpmConfig,
-        writeFile,
-        configDir,
-        loginRegistry,
-        {
-          email,
-          alwaysAuth,
-          _auth,
-        } satisfies BasicAuth
-      ).promise;
+      const result = await saveAuthToUpmConfig(configDir, loginRegistry, {
+        email,
+        alwaysAuth,
+        _auth,
+      } satisfies BasicAuth).promise;
       if (result.isErr()) return result;
       log.notice("config", "saved unity config at " + result.value);
     } else {
@@ -133,17 +120,11 @@ export function makeLoginCmd(
         log.notice("config", `saved to npm config: ${configPath}`)
       );
 
-      const storeResult = await tryStoreUpmAuth(
-        loadUpmConfig,
-        writeFile,
-        configDir,
-        loginRegistry,
-        {
-          email,
-          alwaysAuth,
-          token,
-        } satisfies TokenAuth
-      ).promise;
+      const storeResult = await saveAuthToUpmConfig(configDir, loginRegistry, {
+        email,
+        alwaysAuth,
+        token,
+      } satisfies TokenAuth).promise;
       if (storeResult.isErr()) return storeResult;
       log.notice("config", "saved unity config at " + storeResult.value);
     }
